AI/rice-disease-detection/requirements-updated.txt
fastapi==0.115.6
uvicorn[standard]==0.33.0
python-multipart==0.0.20
ultralytics==8.3.64
roboflow==1.1.50
opencv-python-headless==4.10.0.84
pillow==11.0.0
numpy>=1.24.0,<2.0.0
torch>=2.0.0
torchvision>=0.15.0
python-dotenv==1.0.1
inference-sdk
